+++ changelog.md
@@ -16,6 +16,7 @@ While all _Major_ releases should be reviewed, our only _large_ releases are **v
 - **[change]** - .money() selector and methods
 - **[fix]** - tokenize phone-numbers with parentheses
 - **[fix]** - replaceWith keep-tag issue
+- **[fix]** - preposition-conjunction tagging
 - **[update]** - dependencies
 
 #### 14.15.1 [May 2026]